+++ feature/detail/src/test/kotlin/com/waffiq/bazz_movies/feature/detail/domain/model/omdb/OMDbDetailsTest.kt
@@ -1,5 +1,7 @@
 package com.waffiq.bazz_movies.feature.detail.domain.model.omdb
 
+import com.waffiq.bazz_movies.feature.detail.testutils.HelperTest.oMDbDetails
+import com.waffiq.bazz_movies.feature.detail.testutils.HelperTest.ratingsItem
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
 import org.junit.Test
@@ -39,42 +41,13 @@ class OMDbDetailsTest {
 
   @Test
   fun createOMDbDetails_withAllValues_shouldSetPropertiesCorrectly() {
-    val ratings = listOf(RatingsItem())
-
-    val item = OMDbDetails(
-      metascore = "85",
-      boxOffice = "$100M",
-      website = "example.com",
-      imdbRating = "8.5",
-      imdbVotes = "100,000",
-      ratings = ratings,
-      runtime = "120 min",
-      language = "English",
-      rated = "PG-13",
-      production = "Studio",
-      released = "2024-01-01",
-      imdbID = "tt1234567",
-      plot = "Plot summary",
-      director = "Director Name",
-      title = "Movie Title",
-      actors = "Actor Names",
-      response = "True",
-      type = "movie",
-      awards = "Award Info",
-      dVD = "2024-06-01",
-      year = "2024",
-      poster = "poster.jpg",
-      country = "USA",
-      genre = "Action",
-      writer = "Writer Name"
-    )
-
+    val item = oMDbDetails
     assertEquals("85", item.metascore)
     assertEquals("$100M", item.boxOffice)
     assertEquals("example.com", item.website)
     assertEquals("8.5", item.imdbRating)
     assertEquals("100,000", item.imdbVotes)
-    assertEquals(ratings, item.ratings)
+    assertEquals(ratingsItem, item.ratings)
     assertEquals("120 min", item.runtime)
     assertEquals("English", item.language)
     assertEquals("PG-13", item.rated)
