+++ feature/detail/src/test/kotlin/com/waffiq/bazz_movies/feature/detail/testutils/BaseAdapterTest.kt
